Refactor of validation module (#8)
* Refactor of validation module

* Removed unsed imports
* Changed imports from relative to module style
* Removed unnecessary files
* Renamed modules, functions and instances following PEP8
* Changed validate function to expect
RequestData and return ResponseData object

* Add Makefile

* Add tests folder

* Make pytest output full stacktrace

* Add resources folder and tiger pcl in h5 format

* Add targets to make all and clean to Makefile

* Add help and OS specific switches to Makefile

* Add pytest dependency to validator

* Add js configs to gitignore

* Add missing depepncies to conda_env_validation.yml

* Change validation.config.js template

Changed the template to start serve.py as a module

* Add startup target to Makefile

* Update README to include validator

* Fix typo in README

* Make response return float

Change serve.py so that the response object
just contains a score field of type float and
not a scores field of type List[float] as each
request is only sending one model anyways.

* Changed prints to take formated string

Changed all prints that took multiple arguments to just take one formated string.

* Set validation as root folder

* Fix imports
* Remove TextTo3D
* Move test _validation inside validation folder

* Fix typo in Makefile
