+++ .gitignore
@@ -14,3 +14,7 @@ __pycache__/
 
 # C extensions
 *.so
+
+# autogenerated config files
+generation.config.js
+validation.config.js
