+++ validation/conda_env_validation.yml
@@ -23,3 +23,6 @@ dependencies:
     - torch==2.1.1
     - transformers==4.36.2
     - open-clip-torch==2.24.0
+    - pytest==8.1.1
+    - bittensor==6.9.3
+
